refactored home page
