+++ renderer/home/components/view-dashboard-projects-detail/ViewDashboardProjectsDetail.tsx
@@ -1,30 +1,6 @@
 import Grid from '@material-ui/core/Grid';
-import { makeStyles } from '@material-ui/core/styles';
 import ShowChartIcon from '@material-ui/icons/ShowChart';
-
-const useStyles = makeStyles({
-  ideDetails: {
-    minHeight: 100,
-    backgroundColor: '#FFFFFF',
-    boxShadow: '0px 3px 6px #00000029',
-    borderRadius: '6px',
-    display: 'flex',
-    fontSize: '20px',
-    justifyContent: 'space-evenly',
-    color: '#0075B3',
-    paddingTop: '1em',
-  },
-  showChartIcon: {
-    fontWeight: 'bold',
-    color: '#4CBDEC',
-  },
-  projectDetails: {
-    display: 'flex',
-    flexDirection: 'column',
-    color: '#0075B3',
-    width: '60%',
-  },
-});
+import { useDashboardDetailsStyles } from './ViewDashboardProjectsDetail.styles';
 
 interface Projects {
   name: string;
@@ -35,7 +11,7 @@ export default function ViewDashboardProjectsDetail(props: {
   title: string;
   total: number;
 }): JSX.Element {
-  const classes = useStyles();
+  const classes = useDashboardDetailsStyles();
   return (
     <Grid item xs={4}>
       <div className={classes.ideDetails}>
