+++ renderer/shared/components/accept-button/accept-button.styles.ts
@@ -1,0 +1,9 @@
+import makeStyles from '@material-ui/core/styles/makeStyles';
+
+export const useAcceptButtonStyles = makeStyles({
+  button: {
+    marginTop: '16px',
+    backgroundColor: '#0075B3',
+    color: '#FFFFFF',
+  },
+});
