+++ renderer/shared/components/spinner/spinner.styles.ts
@@ -1,0 +1,9 @@
+import makeStyles from '@material-ui/core/styles/makeStyles';
+
+export const useSpinnerStyles = makeStyles({
+  spinner: {
+    verticalAlign: 'middle',
+    marginTop: '16px',
+    marginLeft: '16px',
+  },
+});
