+++ renderer/shared/hoc/SpaceAround.tsx
@@ -1,6 +1,5 @@
 const styles: { [key: string]: string } = {
   padding: '2rem 0 2rem 1rem',
-  backgroundColor: '#4CBDEC',
 };
 
 interface SpaceAroundProps {
